+++ triplea/cli/next.py
@@ -1,5 +1,5 @@
 import click
-from triplea.service.repository.general import move_state_forward
+from triplea.service.repository.pipeline_core import move_state_forward
 from triplea.service.click_logger import logger
 from triplea.service.repository.persist import refresh
 from triplea.cli.main import cli